Always force alpha=rank and don't allow the user to set it
This simplifies things a bit and makes the LoRA more likely to work with
downstream inference code. Flux uses Diffusers lora saving, which doesn't
include alpha in the state dict, relying on it being in the config file
instead. But ComfyUI can load those state dicts, but the scaling will be
all wrong if alpha != rank.