Fix bug causing first part of dataset to be consistently skipped with resuming from checkpoint
Bug was introduced in 5c1b48649654ae71e064b00ef6310629973b0591. We no longer recreate
the dataloader each iteration. But when resuming from checkpoint, we make a dataloader
that skips the first n batches (because we checkpointed part way through an epoch). So
now, when resuming, we have to make sure to recreate the dataloader just once after the
first iteration so it only skips some examples once.