Remove assert
For video training, it's possible for all videos within a size bucket to
be dropped because none of them are at least as long as the shortest
frame bucket. If that happens this assert will be triggered. But it's
okay to have a ConcatenatedBatchedDataset of size 0, so remove the
assert so training can proceed.