Fix HunyuanVideo VAE tiling sample size
I had copied the config file from Kijai's HunyuanVideoWrapper. But that
codebase changed the name of this config option, and included code changes
to make it work. When using the config file with the official upstream
code, the sample_size incorrectly defaults to 32 instead of 256. This
causes the loss to be slightly higher than it should be.