Fix SDXL pipeline parallelism and switch to non-reentrant gradient checkpointing
After the SDXL pipeline parallelism fix, I was getting errors like
"element 1 of tensors does not require grad and does not have a grad_fn".
Even trying to force a bunch of tensors to require grad in places didn't
fix it. But switching to non-reentrant activation checkpoint worked.
Non-reentrant is the recommended option in Pytorch, but not the default
in deepspeed. I tested all supported models and they are able to train.
Hopefully it's safe to make this change globally.